feat: proper monorepo setup
used pnpm workspaces for a robust monorepo setup.

in this monorepo we can build shared packages which enables us to share
code between frontend and backend easily.

this also enables the sharing of api routes along with input and output
types directly to the frontend jut by writing a proper controller with
input validation and proper return types